Add a common base class for generative models
More to come here... we could consider a default generate step, a
separate subclass for seq2seq.

For now we just pull out the code that is shared across all generative
models and establish the class.
